put NexMarkGoogleRunner.monitor in upper class as of all the linked elements (options, private methods, constants) and remove adherences to GoogleDataFlow i n NexMarkRunner, cleaning issue #28
